put helper funciotn on separated file
